fix(input/erase): warning (conversion)

+++ src/input/specials_characters/erase.c
@@ -10,7 +10,7 @@
 
 void erase(input_t *input)
 {
-    if (input->buffer_size > 0 && ((int)input->cursor_position)
+    if (input->buffer_size > 0 && (input->cursor_position)
     < input->buffer_size) {
         my_putstr("\033[C");
         input->cursor_position++;
